#9 - entity 수정
- 활용 데이터 변경으로 인한 기존 entity 파일 삭제
