--- src/main/java/com/mini/money/repository/DepositDetailRepository.java
@@ -1,7 +1,0 @@
-//package com.mini.money.repository;
-//
-//import com.mini.money.entity.DepositDetail;
-//import org.springframework.data.jpa.repository.JpaRepository;
-//
-//public interface DepositDetailRepository extends JpaRepository<DepositDetail, Long> {
-//}
