--- src/main/java/com/mini/money/repository/DepositRepository.java
@@ -1,7 +1,0 @@
-//package com.mini.money.repository;
-//
-//import com.mini.money.entity.Deposit;
-//import org.springframework.data.jpa.repository.JpaRepository;
-//
-//public interface DepositRepository extends JpaRepository<Deposit, Long> {
-//}
