--- src/main/java/com/mini/money/repository/LoanRepository.java
@@ -1,7 +1,0 @@
-//package com.mini.money.repository;
-//
-//import com.mini.money.entity.Loan;
-//import org.springframework.data.jpa.repository.JpaRepository;
-//
-//public interface LoanRepository extends JpaRepository<Loan, Long> {
-//}
