--- src/main/java/com/mini/money/repository/PublicLoanRepository.java
@@ -1,7 +1,0 @@
-package com.mini.money.repository;
-
-import com.mini.money.entity.PublicLoan;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PublicLoanRepository extends JpaRepository<PublicLoan, Long> {
-}
