--- src/test/java/com/mini/money/api/DepositOptionApiTest.java
@@ -1,22 +1,0 @@
-//package com.mini.money.api;
-//
-//import org.json.simple.parser.ParseException;
-//import org.junit.jupiter.api.Test;
-//import org.springframework.beans.factory.annotation.Autowired;
-//import org.springframework.boot.test.context.SpringBootTest;
-//
-//import java.io.IOException;
-//
-//import static org.junit.jupiter.api.Assertions.*;
-//
-//@SpringBootTest
-//class DepositOptionApiTest {
-//
-//    @Autowired
-//    DepositOptionApi depositOptionApi;
-//
-//    @Test
-//    void name() throws IOException, ParseException {
-//        depositOptionApi.exampleOptionApi();
-//    }
-//}
