+++ src/test/java/com/mini/money/api/LoanApiTest.java
@@ -1,23 +1,23 @@
-//package com.mini.money.api;
-//
-//import org.json.simple.parser.ParseException;
-//import org.junit.jupiter.api.Test;
-//import org.springframework.beans.factory.annotation.Autowired;
-//import org.springframework.boot.test.context.SpringBootTest;
-//
-//import java.io.IOException;
-//
-//import static org.junit.jupiter.api.Assertions.*;
-//
-//@SpringBootTest
-//class LoanApiTest {
-//
-//    @Autowired
-//    HomeLoanApi homeLoanApi;
-//
-//    @Test
-//    void name() throws IOException, ParseException {
-//        homeLoanApi.HomeLoanApi();
-//
-//    }
-//}
+package com.mini.money.api;
+
+import org.json.simple.parser.ParseException;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+
+import java.io.IOException;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+@SpringBootTest
+class PublicLoanApiTest {
+
+    @Autowired
+    PublicLoanApi publicLoanApi;
+
+    @Test
+    void name() throws IOException, ParseException {
+        publicLoanApi.exampleApi();
+
+    }
+}
